+++ examples/index.js
@@ -1,22 +1,25 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
+import Localization from '@/localize'
 
 // Remplace path with build
-import Villain from '../src/index.js'
-import '../src/css/styles.css'
+import Villain from '@/index.js'
+import '@/css/styles.css'
 
 const villainOpts = {
   workerUrl: '/build/worker-bundle.js',
+  // Configurations
+  mangaMode: false,
   allowFullScreen: true,
   autoHideControls: false,
-  preview: 4,
+  allowGlobalShortcuts: false,
 }
 
 const testFile = '/build/testFile/Example-archive.zip'
 
 ReactDOM.render(
   <div>
-    <h1> Test example! </h1>
+    <h1> {Localization['example.header']} </h1>
     <Villain file={testFile} options={villainOpts} />
   </div>,
   document.getElementById('app')