+++ src/index.js
@@ -15,6 +15,7 @@ const defaultOpts = {
   workerUrl: null,
   allowFullScreen: true,
   autoHideControls: false,
+  allowGlobalShortcuts: false,
 }
 
 class Villain extends Component {
@@ -39,10 +40,12 @@ class Villain extends Component {
 
     // Set default value in context state
     // Nore: unsure about this, probably not a good idea, please fix it!
-    const { autoHideControls, allowFullScreen } = opts
+    const { autoHideControls, allowFullScreen, mangaMode, allowGlobalShortcuts } = opts
 
     return (
-      <ReaderProvider defaultState={{ autoHideControls }}>
+      <ReaderProvider
+        defaultState={{ autoHideControls, mangaMode, allowGlobalShortcuts }}
+      >
         <Wrapp width={width} height={height} preview={opts.preview}>
           {container => (
             <Uncompress file={file} workerUrl={opts.workerUrl} preview={opts.preview}>
@@ -54,6 +57,8 @@ class Villain extends Component {
                     focus={state.focus}
                     container={container}
                     bookMode={state.bookMode}
+                    mangaMode={state.mangaMode}
+                    allowGlobalShortcuts={state.allowGlobalShortcuts}
                     currentPage={state.currentPage}
                     allowFullScreen={allowFullScreen}
                     autoHideControls={state.autoHideControls}